// src/api/generated/createContentRequestType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Arannati API
 * API gateway for Arannati microservices
 * OpenAPI spec version: 1.0
 */

export type CreateContentRequestType =
	(typeof CreateContentRequestType)[keyof typeof CreateContentRequestType]

export const CreateContentRequestType = {
	VIDEO: 'VIDEO',
	ARTICLE: 'ARTICLE',
	LINK: 'LINK',
} as const